Remove unused imports; pretty-print msgpack-parser output.
We now use the Text.Groom package to pretty-print the output from "show" when
using the msgpack-parser tool. This helps a human understand the output better,
especially when the output becomes larger.

This commit also removes all unused imports and improves compatibility with
older versions of GHC (7.6). We now use lists instead of vectors in the msgpack
data model.