Record all test inputs, not just the last one per method.
The inputs vary wildly for binary encode/decode methods, so in order to reach
all paths, we now save all 1600+ inputs. This does create a lot of useless
input, because most of them (about 95-99%) are the same. Reducing the maximum
number of passes in quickcheck to 1 solves this problem. Running `make fuzz` in
the toplevel source directory will do this.

Also: fix small bug in binary_encode NodeInfo test. This is only needed for
fuzzing, because the actual test doesn't produce invalid node infos.