Update docker images to new location; add x86.
We now do both arm and x86 android builds on Travis. They don't do much, yet,
just building the library. In the future, we'll try to link them against a C
program that could be executed on a device. We'll then run the x86 version on
an emulator. We won't be doing this anytime soon, though.

For now, this just builds on two different Android architectures and runs no
tests.