Add a short delay before the toxcore test run.
The test suite can sometimes start too fast while the sut has not
started up, yet. This adds a short delay to avoid the issue. The real
fix would be to try connecting until it's up, and then start the test
suite. That is a bit too much work for (probably) no gain, so we're
using a delay instead.