Remove RPC testing framework; use regular local tests.
We have all sorts of problems because of this smart stuff. The idea
behind it was that other implementations (e.g. c-toxcore) could implement
an alternative SUT backend. This never happened. Maybe in the future,
we'll build a spec test like that and apply it to hs-toxcore and
c-toxcore at the same time.

Also, build libsodium from source again, since xenial has 1.0.8 which is
too old for Saltine.